fix bug

+++ dpdk-latency/main.c
@@ -185,7 +185,7 @@ inline static unsigned long long timestamp_nanosecs()
 {
     struct timespec timestamp;
     clock_gettime(CLOCK_MONOTONIC, &timestamp);
-    return (CLOCK_PRECISION * timestamp.tv_sec) + timestamp.tv_nsec;
+    return (unsigned long long)(CLOCK_PRECISION * timestamp.tv_sec) + timestamp.tv_nsec;
 }
 
 static void
